Add eups, pybind11 binding, and kernel minimum height
This adds support for building with eups and adds a pybind11
python binding. Additionally, it adds an optional argument to
the run_kht function, allowing an absolute minimum value cut
on the kernel height, in addition to the already available cut
at a fraction of the maximum kernel height found.